requirements.txt
discord.py[voice]
requests==2.24.0
pip
